enable more unit test
