+++ sql/hive/src/test/scala/org/apache/spark/sql/hive/execution/SQLQuerySuite.scala
@@ -17,6 +17,8 @@
 
 package org.apache.spark.sql.hive.execution
 
+import org.scalatest.BeforeAndAfter
+
 import org.apache.spark.sql.catalyst.DefaultParserDialect
 import org.apache.spark.sql.catalyst.analysis.EliminateSubQueries
 import org.apache.spark.sql.catalyst.errors.DialectException
@@ -28,6 +30,7 @@ import org.apache.spark.sql.hive.{HiveQLDialect, MetastoreRelation}
 import org.apache.spark.sql.parquet.ParquetRelation2
 import org.apache.spark.sql.sources.LogicalRelation
 import org.apache.spark.sql.types._
+import org.apache.spark.sql.{AnalysisException, QueryTest, Row, SQLConf}
 
 case class Nested1(f1: Nested2)
 case class Nested2(f2: Nested3)
@@ -903,5 +906,15 @@ class SQLQuerySuite extends QueryTest {
       sql("drop table if exists dynparttest2")
       sql("set hive.exec.dynamic.partition.mode=strict")
     }
+  }
+}
+
+class SQLQuerySuite2 extends SQLQuerySuite with BeforeAndAfter {
+  def beforeAll() {
+    TestHive.setConf(SQLConf.AGGREGATE_2, "true")
+  }
+
+  def afterAll() {
+    TestHive.setConf(SQLConf.AGGREGATE_2, "false")
   }
 }
